Migrate prefix filters over to data source
